Move component identity into separate file
